Optimize black_box
The previous assembly code required the compiler to move the value into memory to be able to pass the pointer into assembly. Often this also resulted in the function not being inlined. I believe this code is equivalent in terms of black box behavior but allows for much better codegen in the case where `input` is already in a register.