Harden stable implementation of black_box
In Rust, `core::ptr::read_volatile` is not formally specified.
However the current documentation mentions that "the semantics
will almost always end up pretty similar to C11's definition of
volatile".

In C11, a volatile read *always* reads a value from memory. So this
value *must* be in RAM in that form, at some time.

This allows us to create a relatively efficient optimization
barrier, by spilling the value to the stack, and immediadiately
reloading it again.

The benefit is that LTO could ever optimize the empty function
away, even if #[inline(always)] is specified (although this may
be very unlikely). However, `volatile` is concretely specified,
so this optimization barrier will not be lost.