Implement `From<CtOption<T>> for Option<T>
As stated in #73 and #74 the `CtOption<T>` API is a bit
unflexible when you try to extract the value from the
`CtOption` or you want to convert the `CtOption<T>`
into an `Option<T>` or a `Result<T>`.

Therefore the conversion from `CtOption<T>` to `Option<T>`
has been implemented to avoid having situations like:
```rust
if ct_opt.is_some() {
     return ct_opt.unwrap()
} else {
    return Err(....)
}
```

 # NOTE:
This impl is done to avoid ending up with unpractical, verbose and/or bad handled
conversions from the `CtOption<T>` wraps to an `Option<T>` or `Result<T, E>`.
This implementation doesn't intend to be constant-time nor try to protect the
leakage of the `T` since the `Option<T>` will do it anyways.

Closes #73 & #74