Change Rust version for MSRV test case
The code uploaded to solve #73 & #74 includes a feature that
was added to the Rust language in Rust_1.41.0 via RFC2451.

Therefore, the minimum version needed in order to allow the tests
pass for the MSRV testcase is the one mentioned avobe.

As suggested per @hdevalence, it might not be worth it to try to
support a +4yo version of Rust, so the .travis.yml has been changed
in order to update the version used to test the MSRV case to 1.41.0.