[WIP] Attempt at making a constant-time PartialOrd impl.
I don't actually like this approach because there's no way to get the
integration with all of Rust's functionality through the
Ordering::{Less,Equal,Greater} enum without branching on whether the value is
-1, 0, or 1.  Because Ordering is implemented internally with i8s:

    pub enum Ordering {
        Less = -1,
        Equal = 0,
        Greater = 1,
    }

I assumed we could just construct an Ordering like so:

    let x = -1;
    let cmp = Ordering(x);

    assert!(::std::matches!(cmp, Ordering::Less {..} ));

But that turns out to not be the case?  If it is, then I don't see a clear way
to make this constant-time.