Add `CtOption::expect`
Along the lines of the existing `CtOption::unwrap`, this adds a
corresponding `CtOption::expect` which works like `Option::expect`:
unwraps the value if it's some, or else panics with a provided error
message.

Using `expect` over `unwrap` is generally prefered from a Rust style
perspective. This allows the caller to describe what invariant was
violated when it's expected some `CtOption` is always some.