Add ConstantTimeEq implementation for [T; N].
Only when the const-generics feature is enabled.

I had to modify a few tests: they were relying on [u8; N] being
automatically dereferenced into &[u8] to get the blanket ConstantTimeEq
implementation for [T]. But once a blanket implementation is also
available for [T; N], Rust attempts to use it instead & then complains
that it can't compare arrays of different lengths. I suppose this
technically counts as a backwards-compatibility break.